Session 4 - IoT Digital Twin/Functions/twin_updater/requirements.txt
azure-functions
azure-digitaltwins-core 
azure-identity